Lambda Native Environment Variables and Removal of Streambot (#42)
* Removing streambot usage from repo and tests

Removed the streambot wrapping of the lambda function
as well as the references to streambot env variables.

Removed streambot specfic unit tests.

Added envVariableParser() to replace the formatter
for streambot and set that to Enviornment.Variables
in the CFN script output.

* Fixing small issues and cleaning up templates

Removing streambot references in templates as well
as adding the newer env variables to the templates

* Removing Streambot references from templates

Fixed a bug where the templating was adding an extra section
from the last commit. Also, removed references to streambot
in the templates to correct the tests.

6 tests still failing due to needing to add the env
variables into the templates

* Modifing templates to match new env layout

In lambda-cfn reverting the removal of the ref parsing
for the envVariableParser. This helped make the unit
tests align better with the expected output

Updating all of the templates to match the
aws lambda sytle env variable style with
Environment.Variables object.

* Changing lambdacfalarm status to aws reference

Changing the LambdaCfnAlarmSNSTopic from static bool
to a reference in lambda-cfn.js

Updating templates to match said changes

* Removing streambot from package.json

Also, changing const to var for consistency

* Updating documentation with removal of streambot

Added reference to AWS Lambda functions for ease of use

* Adding unit tests for envVariableParser()

Also adding envVariableParser to the exported functions
to be able to test/use elsewhere

* Changing the names of new unit tests

Replacing streambot with envVariableParser()

* Updating CHANGELOG.md for 1.0.0